requirements.txt

bcrypt==3.2.0
cryptography==3.4.8
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
jsonschema==3.2.0
PyJWT==2.3.0
SQLAlchemy==2.0.28
gunicorn==19.7.1



